practica_logstash/src/main/webapp/app/entities/mensaje-privado/route/mensaje-privado-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IMensajePrivado } from '../mensaje-privado.model';
import { MensajePrivadoService } from '../service/mensaje-privado.service';

import { MensajePrivadoRoutingResolveService } from './mensaje-privado-routing-resolve.service';

describe('MensajePrivado routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: MensajePrivadoRoutingResolveService;
  let service: MensajePrivadoService;
  let resultMensajePrivado: IMensajePrivado | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(MensajePrivadoRoutingResolveService);
    service = TestBed.inject(MensajePrivadoService);
    resultMensajePrivado = undefined;
  });

  describe('resolve', () => {
    it('should return IMensajePrivado returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMensajePrivado = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMensajePrivado).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMensajePrivado = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultMensajePrivado).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IMensajePrivado>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMensajePrivado = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMensajePrivado).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
